+++ Sources/KeyboardKit/Layout/KeyboardLayout.swift
@@ -9,8 +9,8 @@
 import CoreGraphics
 import SwiftUI
 
-/// A keyboard layout contains the full set of keyboard keys
-/// that should be rendered for a keyboard.
+/// This type defines the full set of keys on a keyboard and
+/// serves as a namespace for layout-related types.
 ///
 /// A layout also specifies sizes, alignments, etc. which is
 /// required information when rendering a keyboard.