+++ Sources/KeyboardKit/Localization/KeyboardLocale.swift
@@ -9,16 +9,13 @@
 import Foundation
 
 /// This enum defines KeyboardKit-supported locales and is a
-/// namespace for related types.
+/// namespace for locale-elated types.
 ///
 /// Every keyboard locale refers to a native ``locale`` that
 /// provide locale-specific information. 
 ///
 /// A keyboard locale also defines localized assets that can
 /// be translated with ``KKL10n``.
-///
-/// The namespace doesn't contain protocols, open classes or
-/// types of higher importance.
 public enum KeyboardLocale: String, CaseIterable, Codable, Identifiable {
     
     /// Try to map a fuzzy name to a keyboard locale.