+++ Sources/KeyboardKit/Settings/KeyboardSettings.swift
@@ -1,0 +1,14 @@
+//
+//  KeyboardSettings.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2024-03-30.
+//  Copyright © 2024 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/// This namespace contains keyboard settings-related types.
+///
+/// The namespace doesn't contain protocols or context types.
+public struct KeyboardSettings {}