+++ Sources/KeyboardKit/_Deprecated/KeyboardSettings+Deprecated.swift
@@ -1,0 +1,4 @@
+import SwiftUI
+
+@available(*, deprecated, renamed: "KeyboardSettings.Link")
+public typealias KeyboardSettingsLink = KeyboardSettings.Link