+++ Sources/KeyboardKit/_Essentials/Keyboard.swift
@@ -10,6 +10,5 @@ import Foundation
 
 /// This namespace contains essential keyboard-related types.
 ///
-/// The namespace doesn't contain protocols, open classes or
-/// types of higher importance.
+/// The namespace doesn't contain protocols or context types.
 public struct Keyboard {}